feat: expose jsToMdxTreeNode util
