+++ packages/integrations/mdx/package.json
@@ -20,6 +20,7 @@
   "homepage": "https://docs.astro.build/en/guides/integrations-guide/mdx/",
   "exports": {
     ".": "./dist/index.js",
+    "./utils": "./dist/utils.js",
     "./package.json": "./package.json"
   },
   "scripts": {
